moved type helper methods into separate files
